requirements.txt
PyMySQL==1.1.0
jmespath==1.0.1
pip==22.3.1
python-dateutil==2.8.2
setuptools==65.5.1
six==1.16.0
urllib3==1.26.16
wheel==0.38.4
PyJWT==2.8.0
